+++ rti_mote/ir_comm.h
@@ -29,7 +29,7 @@ class ir_comm;
 
 class ir_comm {
  private:
-  byte* p_write;
+  int* p_write;
 
  public:
   /**
@@ -43,6 +43,6 @@ class ir_comm {
   void send();
   float rss();
   void receive();
-  void set_p_write(byte* irRSS);
+  void set_p_write(int* irRSS);
 };
 #endif /*IR_COMM_H*/
